+++ src/typeorm/entities/StaffRole.ts
@@ -2,7 +2,7 @@ import { Column, Entity, Index, PrimaryGeneratedColumn } from 'typeorm';
 
 @Entity({ name: 'staff_roles' })
 @Index(['guildId'])
-export class SavedRole {
+export class StaffRole {
   @PrimaryGeneratedColumn()
   id: number;